refactor: schedule page
